change CardTile component to ts
