--- client/src/components/profile/deckEditor/CardTile.js
@@ -1,17 +1,0 @@
-import React from "react";
-
-const CardTile = (props) => {
-  const { index, card, cards, setCards } = props;
-
-  const removeCard = () => {
-    cards.splice(index, 1);
-    setCards([...cards]);
-  };
-  return (
-    <div className="card-tile" onClick={removeCard}>
-      {card.name}
-    </div>
-  );
-};
-
-export default CardTile;
